fix: change header realization
There is an incorrect nesting of HTML elements in the header which leads
to an error during the validation process. <ul> tag can't be a
parent tag for <div> child element.

In order to save the primary code structure the <ul> element was
replaced by an usual <div>. All <li> elements were removed; theirs only
css property was added to .navigation__link selector as it's stated in
the diff report.